Микросервис | Fiber [#505569 #559645]

+++ internal/api/delete_item.go
@@ -7,11 +7,13 @@ import (
 
 func (s *Server) DeleteItem(c *fiber.Ctx) error {
 	id := c.Params("id")
+	if id == "" {
+		return fiber.NewError(fiber.StatusBadRequest, "id is required")
+	}
 	err := s.Repository.Delete(c.Context(), id)
 	if err != nil {
 		log.Errorw("s.Repository.Delete", err)
 		return fiber.NewError(fiber.StatusInternalServerError, "internal server error")
 	}
 	return c.Status(fiber.StatusNoContent).Send(nil)
-	//return c.Status(fiber.StatusCreated).JSON(fiber.Map{})
 }
